Plasmid Plotting Function

Plotting function to create the final plasmid map from properly file.

plasmid_plot(
  arrow_df,
  labels,
  angle_adjustment,
  zoom_y,
  plasmid_name,
  font_family = "mono",
  name_size,
  bp_count = TRUE,
  curved_size,
  curve,
  label_nudge,
  label_size,
  repel_box,
  plasmid_length,
  plasmid_y = 2,
  plasmid_width = 0.1
)

Arguments

arrow_df

Dataframe of points from the create_arrow_plotting_df() function to draw arrows.

labels

List resulting from create_labels() function containing curved text dataframe and labels for annotation.

angle_adjustment

Angle (in degrees) to rotate the resulting plasmid map by.

zoom_y

Zoom level for the resulting plasmid map. Sets ylim(0,zoom_y) and can be overriden.

plasmid_name

String name for the plasmid map.

font_family

Font family to use in all of the text on the plot.

name_size

Size of the plasmid name annotation.

curved_size

Font size of the curved labels (in .pt).

curve

Curvature of the repelled text labels.

label_nudge

Amount to nudge the labels away from the plasmid arrows to allow for neater packing.

label_size

Size of the repelled labels.

repel_box

Distance to repel between the labels.

plasmid_length

Total length of the plasmid.

plasmid_y

Position on the y axis of the original plasmid.

plasmid_width

Width of the arrows of the plasmid.