Read a .gb GenBank File

Read a .gb GenBank File

read_gb(file)

Arguments

file

File path or connection, that can be handled by readr::read_lines().

Value

a list of class 'plasmid' which can be further coerced for plotting.

Examples

fl <- system.file("extdata", "petm20.gb", package = "plasmapR")

fl |>
  read_gb() |>
  as.data.frame()
#> Warning: row names were found from a short variable and have been discarded
#> Warning: row names were found from a short variable and have been discarded
#> Warning: row names were found from a short variable and have been discarded
#> Warning: row names were found from a short variable and have been discarded
#>   length features.type features.index features.start_end features.direction
#> 1   7700        source              1                  1                  1
#> 2   7700        source              1               7700                  1
#>         features.organism features.mol_type           features.name
#> 1 synthetic DNA construct         other DNA synthetic DNA construct
#> 2 synthetic DNA construct         other DNA synthetic DNA construct
#>   features.type.1 features.index.1 features.start_end.1 features.direction.1
#> 1      rep_origin                2                   12                    1
#> 2      rep_origin                2                  467                    1
#>   features.label
#> 1        f1 orim
#> 2        f1 orim
#>                                                                               features.note
#> 1 f1 bacteriophage origin of replication; arrow indicates direction of (+) strand synthesis
#> 2 f1 bacteriophage origin of replication; arrow indicates direction of (+) strand synthesis
#>   features.name.1 features.type.2 features.index.2 features.start_end.2
#> 1         f1 orim        promoter                3                  494
#> 2         f1 orim        promoter                3                  598
#>   features.direction.2 features.gene features.label.1 features.name.2
#> 1                    1           bla    AmpR promoter   AmpR promoter
#> 2                    1           bla    AmpR promoter   AmpR promoter
#>   features.type.3 features.index.3 features.start_end.3 features.direction.3
#> 1             CDS                4                  599                    1
#> 2             CDS                4                 1459                    1
#>   features.codon_start features.gene.1 features.product features.label.2
#> 1                    1             bla   beta-lactamase             AmpR
#> 2                    1             bla   beta-lactamase             AmpR
#>                                                           features.note.1
#> 1 confers resistance to ampicillin, carbenicillin, andrelated antibiotics
#> 2 confers resistance to ampicillin, carbenicillin, andrelated antibiotics
#>                                                                                                                                                                                                                                                                             features.translation
#> 1 MSIQHFRVALIPFFAAFCLPVFAHPETLVKVKDAEDQLGARVGYIELDLNSGKILESFRPEERFPMMSTFKVLLCGAVLSRIDAGQEQLGRRIHYSQNDLVEYSPVTEKHLTDGMTVRELCSAAITMSDNTAANLLLTTIGGPKELTAFLHNMGDHVTRLDRWEPELNEAIPNDERDTTMPAAMATTLRKLLTGELLTLASRQQLIDWMEADKVAGPLLRSALPAGWFIADKSGAGERGSRGIIAALGPDGKPSRIVVIYTTGSQATMDERNRQIAEIGASLIKHW
#> 2 MSIQHFRVALIPFFAAFCLPVFAHPETLVKVKDAEDQLGARVGYIELDLNSGKILESFRPEERFPMMSTFKVLLCGAVLSRIDAGQEQLGRRIHYSQNDLVEYSPVTEKHLTDGMTVRELCSAAITMSDNTAANLLLTTIGGPKELTAFLHNMGDHVTRLDRWEPELNEAIPNDERDTTMPAAMATTLRKLLTGELLTLASRQQLIDWMEADKVAGPLLRSALPAGWFIADKSGAGERGSRGIIAALGPDGKPSRIVVIYTTGSQATMDERNRQIAEIGASLIKHW
#>   features.name.3 features.type.4 features.index.4 features.start_end.4
#> 1            AmpR      rep_origin                5                 1630
#> 2            AmpR      rep_origin                5                 2218
#>   features.direction.4 features.label.3
#> 1                    1              ori
#> 2                    1              ori
#>                                                features.note.2 features.name.4
#> 1 high-copy-number ColE1/pMB1/pBR322/pUC origin of replication             ori
#> 2 high-copy-number ColE1/pMB1/pBR322/pUC origin of replication             ori
#>   features.type.5 features.index.5 features.start_end.5 features.direction.5
#> 1    misc_feature                6                 2404                    1
#> 2    misc_feature                6                 2546                    1
#>   features.label.4                      features.note.3 features.name.5
#> 1              bom basis of mobility region from pBR322             bom
#> 2              bom basis of mobility region from pBR322             bom
#>   features.type.6 features.index.6 features.start_end.6 features.direction.6
#> 1             CDS                7                 2648                   -1
#> 2             CDS                7                 2839                   -1
#>   features.codon_start.1 features.gene.2
#> 1                      1             rop
#> 2                      1             rop
#>                                        features.product.1 features.label.5
#> 1 Rop protein, which maintains plasmids at low copynumber              rop
#> 2 Rop protein, which maintains plasmids at low copynumber              rop
#>                                            features.translation.1
#> 1 MTKQEKTALNMARFIRSQTLTLLEKLNELDADEQADICESLHDHADELYRSCLARFGDDGENL
#> 2 MTKQEKTALNMARFIRSQTLTLLEKLNELDADEQADICESLHDHADELYRSCLARFGDDGENL
#>   features.name.6 features.type.7 features.index.7 features.start_end.7
#> 1             rop             CDS                8                 3648
#> 2             rop             CDS                8                 4730
#>   features.direction.7 features.codon_start.2 features.gene.3
#> 1                    1                      1            lacI
#> 2                    1                      1            lacI
#>   features.product.2 features.label.6
#> 1      lac repressor             lacI
#> 2      lac repressor             lacI
#>                                                                                                                                                                        features.note.4
#> 1 The lac repressor binds to the lac operator to inhibit transcription in E. coli. This inhibition can be relieved by adding lactose or isopropyl-beta-D-thiogalactopyranoside (IPTG).
#> 2 The lac repressor binds to the lac operator to inhibit transcription in E. coli. This inhibition can be relieved by adding lactose or isopropyl-beta-D-thiogalactopyranoside (IPTG).
#>                                                                                                                                                                                                                                                                                                                                                     features.translation.2
#> 1 MKPVTLYDVAEYAGVSYQTVSRVVNQASHVSAKTREKVEAAMAELNYIPNRVAQQLAGKQSLLIGVATSSLALHAPSQIVAAIKSRADQLGASVVVSMVERSGVEACKAAVHNLLAQRVSGLIINYPLDDQDAIAVEAACTNVPALFLDVSDQTPINSIIFSHEDGTRLGVEHLVALGHQQIALLAGPLSSVSARLRLAGWHKYLTRNQIQPIAEREGDWSAMSGFQQTMQMLNEGIVPTAMLVANDQMALGAMRAITESGLRVGADISVVGYDDTEDSSCYIPPLTTIKQDFRLLGQTSVDRLLQLSQGQAVKGNQLLPVSLVKRKTTLAPNTQTASPRALADSLMQLARQVSRLESGQ
#> 2 MKPVTLYDVAEYAGVSYQTVSRVVNQASHVSAKTREKVEAAMAELNYIPNRVAQQLAGKQSLLIGVATSSLALHAPSQIVAAIKSRADQLGASVVVSMVERSGVEACKAAVHNLLAQRVSGLIINYPLDDQDAIAVEAACTNVPALFLDVSDQTPINSIIFSHEDGTRLGVEHLVALGHQQIALLAGPLSSVSARLRLAGWHKYLTRNQIQPIAEREGDWSAMSGFQQTMQMLNEGIVPTAMLVANDQMALGAMRAITESGLRVGADISVVGYDDTEDSSCYIPPLTTIKQDFRLLGQTSVDRLLQLSQGQAVKGNQLLPVSLVKRKTTLAPNTQTASPRALADSLMQLARQVSRLESGQ
#>   features.name.7 features.type.8 features.index.8 features.start_end.8
#> 1            lacI        promoter                9                 4731
#> 2            lacI        promoter                9                 4808
#>   features.direction.8 features.gene.4 features.label.7 features.name.8
#> 1                    1            lacI    lacI promoter   lacI promoter
#> 2                    1            lacI    lacI promoter   lacI promoter
#>   features.type.9 features.index.9 features.start_end.9 features.direction.9
#> 1        promoter               10                 5121                    1
#> 2        promoter               10                 5139                    1
#>   features.label.8                              features.note.5 features.name.9
#> 1      T7 promoter promoter for bacteriophage T7 RNA polymerase     T7 promoter
#> 2      T7 promoter promoter for bacteriophage T7 RNA polymerase     T7 promoter
#>   features.type.10 features.index.10 features.start_end.10
#> 1     protein_bind                11                  5140
#> 2     protein_bind                11                  5164
#>   features.direction.10 features.label.9         features.bound_moiety
#> 1                     1     lac operator lac repressor encoded by lacI
#> 2                     1     lac operator lac repressor encoded by lacI
#>                                                                                                                                                                        features.note.6
#> 1 The lac repressor binds to the lac operator to inhibit transcription in E. coli. This inhibition can be relieved by adding lactose or isopropyl-beta-D-thiogalactopyranoside (IPTG).
#> 2 The lac repressor binds to the lac operator to inhibit transcription in E. coli. This inhibition can be relieved by adding lactose or isopropyl-beta-D-thiogalactopyranoside (IPTG).
#>   features.name.10 features.type.11 features.index.11 features.start_end.11
#> 1     lac operator              CDS                12                  5209
#> 2     lac operator              CDS                12                  5535
#>   features.direction.11 features.codon_start.3 features.gene.5
#> 1                     1                      1            trxA
#> 2                     1                      1            trxA
#>    features.product.3 features.label.10
#> 1 E. coli thioredoxin              TrxA
#> 2 E. coli thioredoxin              TrxA
#>                                                                                          features.translation.3
#> 1 MSDKIIHLTDDSFDTDVLKADGAILVDFWAEWCGPCKMIAPILDEIADEYQGKLTVAKLNIDQNPGTAPKYGIRGIPTLLLFKNGEVAATKVGALSKGQLKEFLDANLA
#> 2 MSDKIIHLTDDSFDTDVLKADGAILVDFWAEWCGPCKMIAPILDEIADEYQGKLTVAKLNIDQNPGTAPKYGIRGIPTLLLFKNGEVAATKVGALSKGQLKEFLDANLA
#>   features.name.11 features.type.12 features.index.12 features.start_end.12
#> 1             TrxA              CDS                13                  5557
#> 2             TrxA              CDS                13                  5574
#>   features.direction.12 features.codon_start.4 features.product.4
#> 1                     1                      1 6xHis affinity tag
#> 2                     1                      1 6xHis affinity tag
#>   features.label.11 features.translation.4 features.name.12 features.type.13
#> 1             6xHis                 HHHHHH            6xHis      primer_bind
#> 2             6xHis                 HHHHHH            6xHis      primer_bind
#>   features.index.13 features.start_end.13 features.direction.13
#> 1                14                  5578                    -1
#> 2                14                  5628                    -1
#>   features.label.12 features.name.13 features.type.14 features.index.14
#> 1   AviTag Insert R  AviTag Insert R              CDS                15
#> 2   AviTag Insert R  AviTag Insert R              CDS                15
#>   features.start_end.14 features.direction.14 features.codon_start.5
#> 1                  5584                     1                      1
#> 2                  5604                     1                      1
#>                                               features.product.5
#> 1 tobacco etch virus (TEV) protease recognition andcleavage site
#> 2 tobacco etch virus (TEV) protease recognition andcleavage site
#>   features.label.13 features.translation.5 features.name.14 features.type.15
#> 1          TEV Site                ENLYFQG         TEV Site              CDS
#> 2          TEV Site                ENLYFQG         TEV Site              CDS
#>   features.index.15 features.start_end.15 features.direction.15
#> 1                16                  5611                     1
#> 2                16                  5655                     1
#>   features.codon_start.6                                  features.product.6
#> 1                      1 peptide tag that allows for enzymatic biotinylation
#> 2                      1 peptide tag that allows for enzymatic biotinylation
#>   features.label.14 features.translation.6 features.name.15 features.type.16
#> 1        AviTag(TM)        GLNDIFEAQKIEWHE       AviTag(TM)      primer_bind
#> 2        AviTag(TM)        GLNDIFEAQKIEWHE       AviTag(TM)      primer_bind
#>   features.index.16 features.start_end.16 features.direction.16
#> 1                17                  5629                     1
#> 2                17                  5673                     1
#>   features.label.15 features.name.16 features.type.17 features.index.17
#> 1   AviTag Insert F  AviTag Insert F              CDS                18
#> 2   AviTag Insert F  AviTag Insert F              CDS                18
#>   features.start_end.17 features.direction.17 features.codon_start.7
#> 1                  5659                     1                      1
#> 2                  7503                     1                      1
#>   features.label.16
#> 1       dsnPPR10-C2
#> 2       dsnPPR10-C2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   features.translation.7
#> 1 MGNDVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYTTLIDGLAKAGRLEEALCLFQEMKEKGVKPNVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPNVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPNVVTYNTLIDGLAKAGRLEEALCLFQEMKEKGVKPSVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPNVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTNNTLKDGASKAG
#> 2 MGNDVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYTTLIDGLAKAGRLEEALCLFQEMKEKGVKPNVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPNVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPNVVTYNTLIDGLAKAGRLEEALCLFQEMKEKGVKPSVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTYTTLIDGLAKAGRLEEALQLFQEMKEKGVKPNVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPDVVTYNTLIDGLAKAGRLEEALQLFQEMKEKGVKPSVVTNNTLKDGASKAG
#>   features.name.17 features.type.18 features.index.18 features.start_end.18
#> 1      dsnPPR10-C2              CDS                19                  7544
#> 2      dsnPPR10-C2              CDS                19                  7561
#>   features.direction.18 features.codon_start.8 features.product.7
#> 1                     1                      1 6xHis affinity tag
#> 2                     1                      1 6xHis affinity tag
#>   features.label.17 features.translation.8 features.name.18 features.type.19
#> 1             6xHis                 HHHHHH            6xHis       terminator
#> 2             6xHis                 HHHHHH            6xHis       terminator
#>   features.index.19 features.start_end.19 features.direction.19
#> 1                20                  7628                     1
#> 2                20                  7675                     1
#>   features.label.18
#> 1     T7 terminator
#> 2     T7 terminator
#>                                                features.note.7 features.name.19
#> 1 transcription terminator for bacteriophage T7 RNA polymerase    T7 terminator
#> 2 transcription terminator for bacteriophage T7 RNA polymerase    T7 terminator
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               sequence
#> 1 tggcgaatgggacgcgccctgtagcggcgcattaagcgcggcgggtgtggtggttacgcgcagcgtgaccgctacacttgccagcgccctagcgcccgctcctttcgctttcttcccttcctttctcgccacgttcgccggctttccccgtcaagctctaaatcgggggctccctttagggttccgatttagtgctttacggcacctcgaccccaaaaaacttgattagggtgatggttcacgtagtgggccatcgccctgatagacggtttttcgccctttgacgttggagtccacgttctttaatagtggactcttgttccaaactggaacaacactcaaccctatctcggtctattcttttgatttataagggattttgccgatttcggcctattggttaaaaaatgagctgatttaacaaaaatttaacgcgaattttaacaaaatattaacgtttacaatttcaggtggcacttttcggggaaatgtgcgcggaacccctatttgtttatttttctaaatacattcaaatatgtatccgctcatgagacaataaccctgataaatgcttcaataatattgaaaaaggaagagtatgagtattcaacatttccgtgtcgcccttattcccttttttgcggcattttgccttcctgtttttgctcacccagaaacgctggtgaaagtaaaagatgctgaagatcagttgggtgcacgagtgggttacatcgaactggatctcaacagcggtaagatccttgagagttttcgccccgaagaacgttttccaatgatgagcacttttaaagttctgctatgtggcgcggtattatcccgtattgacgccgggcaagagcaactcggtcgccgcatacactattctcagaatgacttggttgagtactcaccagtcacagaaaagcatcttacggatggcatgacagtaagagaattatgcagtgctgccataaccatgagtgataacactgcggccaacttacttctgacaacgatcggaggaccgaaggagctaaccgcttttttgcacaacatgggggatcatgtaactcgccttgatcgttgggaaccggagctgaatgaagccataccaaacgacgagcgtgacaccacgatgcctgcagcaatggcaacaacgttgcgcaaactattaactggcgaactacttactctagcttcccggcaacaattaatagactggatggaggcggataaagttgcaggaccacttctgcgctcggcccttccggctggctggtttattgctgataaatctggagccggtgagcgtgggtctcgcggtatcattgcagcactggggccagatggtaagccctcccgtatcgtagttatctacacgacggggagtcaggcaactatggatgaacgaaatagacagatcgctgagataggtgcctcactgattaagcattggtaactgtcagaccaagtttactcatatatactttagattgatttaaaacttcatttttaatttaaaaggatctaggtgaagatcctttttgataatctcatgaccaaaatcccttaacgtgagttttcgttccactgagcgtcagaccccgtagaaaagatcaaaggatcttcttgagatcctttttttctgcgcgtaatctgctgcttgcaaacaaaaaaaccaccgctaccagcggtggtttgtttgccggatcaagagctaccaactctttttccgaaggtaactggcttcagcagagcgcagataccaaatactgtccttctagtgtagccgtagttaggccaccacttcaagaactctgtagcaccgcctacatacctcgctctgctaatcctgttaccagtggctgctgccagtggcgataagtcgtgtcttaccgggttggactcaagacgatagttaccggataaggcgcagcggtcgggctgaacggggggttcgtgcacacagcccagcttggagcgaacgacctacaccgaactgagatacctacagcgtgagctatgagaaagcgccacgcttcccgaagggagaaaggcggacaggtatccggtaagcggcagggtcggaacaggagagcgcacgagggagcttccagggggaaacgcctggtatctttatagtcctgtcgggtttcgccacctctgacttgagcgtcgatttttgtgatgctcgtcaggggggcggagcctatggaaaaacgccagcaacgcggcctttttacggttcctggccttttgctggccttttgctcacatgttctttcctgcgttatcccctgattctgtggataaccgtattaccgcctttgagtgagctgataccgctcgccgcagccgaacgaccgagcgcagcgagtcagtgagcgaggaagcggaagagcgcctgatgcggtattttctccttacgcatctgtgcggtatttcacaccgcatatatggtgcactctcagtacaatctgctctgatgccgcatagttaagccagtatacactccgctatcgctacgtgactgggtcatggctgcgccccgacacccgccaacacccgctgacgcgccctgacgggcttgtctgctcccggcatccgcttacagacaagctgtgaccgtctccgggagctgcatgtgtcagaggttttcaccgtcatcaccgaaacgcgcgaggcagctgcggtaaagctcatcagcgtggtcgtgaagcgattcacagatgtctgcctgttcatccgcgtccagctcgttgagtttctccagaagcgttaatgtctggcttctgataaagcgggccatgttaagggcggttttttcctgtttggtcactgatgcctccgtgtaagggggatttctgttcatgggggtaatgataccgatgaaacgagagaggatgctcacgatacgggttactgatgatgaacatgcccggttactggaacgttgtgagggtaaacaactggcggtatggatgcggcgggaccagagaaaaatcactcagggtcaatgccagcgcttcgttaatacagatgtaggtgttccacagggtagccagcagcatcctgcgatgcagatccggaacataatggtgcagggcgctgacttccgcgtttccagactttacgaaacacggaaaccgaagaccattcatgttgttgctcaggtcgcagacgttttgcagcagcagtcgcttcacgttcgctcgcgtatcggtgattcattctgctaaccagtaaggcaaccccgccagcctagccgggtcctcaacgacaggagcacgatcatgcgcacccgtggggccgccatgccggcgataatggcctgcttctcgccgaaacgtttggtggcgggaccagtgacgaaggcttgagcgagggcgtgcaagattccgaataccgcaagcgacaggccgatcatcgtcgcgctccagcgaaagcggtcctcgccgaaaatgacccagagcgctgccggcacctgtcctacgagttgcatgataaagaagacagtcataagtgcggcgacgatagtcatgccccgcgcccaccggaaggagctgactgggttgaaggctctcaagggcatcggtcgagatcccggtgcctaatgagtgagctaacttacattaattgcgttgcgctcactgcccgctttccagtcgggaaacctgtcgtgccagctgcattaatgaatcggccaacgcgcggggagaggcggtttgcgtattgggcgccagggtggtttttcttttcaccagtgagacgggcaacagctgattgcccttcaccgcctggccctgagagagttgcagcaagcggtccacgctggtttgccccagcaggcgaaaatcctgtttgatggtggttaacggcgggatataacatgagctgtcttcggtatcgtcgtatcccactaccgagatgtccgcaccaacgcgcagcccggactcggtaatggcgcgcattgcgcccagcgccatctgatcgttggcaaccagcatcgcagtgggaacgatgccctcattcagcatttgcatggtttgttgaaaaccggacatggcactccagtcgccttcccgttccgctatcggctgaatttgattgcgagtgagatatttatgccagccagccagacgcagacgcgccgagacagaacttaatgggcccgctaacagcgcgatttgctggtgacccaatgcgaccagatgctccacgcccagtcgcgtaccgtcttcatgggagaaaataatactgttgatgggtgtctggtcagagacatcaagaaataacgccggaacattagtgcaggcagcttccacagcaatggcatcctggtcatccagcggatagttaatgatcagcccactgacgcgttgcgcgagaagattgtgcaccgccgctttacaggcttcgacgccgcttcgttctaccatcgacaccaccacgctggcacccagttgatcggcgcgagatttaatcgccgcgacaatttgcgacggcgcgtgcagggccagactggaggtggcaacgccaatcagcaacgactgtttgcccgccagttgttgtgccacgcggttgggaatgtaattcagctccgccatcgccgcttccactttttcccgcgttttcgcagaaacgtggctggcctggttcaccacgcgggaaacggtctgataagagacaccggcatactctgcgacatcgtataacgttactggtttcacattcaccaccctgaattgactctcttccgggcgctatcatgccataccgcgaaaggttttgcgccattcgatggtgtccgggatctcgacgctctcccttatgcgactcctgcattaggaagcagcccagtagtaggttgaggccgttgagcaccgccgccgcaaggaatggtgcatgcaaggagatggcgcccaacagtcccccggccacggggcctgccaccatacccacgccgaaacaagcgctcatgagcccgaagtggcgagcccgatcttccccatcggtgatgtcggcgatataggcgccagcaaccgcacctgtggcgccggtgatgccggccacgatgcgtccggcgtagaggatcgagatcgatctcgatcccgcgaaattaatacgactcactataggggaattgtgagcggataacaattcccctctagaaataattttgtttaactttaagaaggagatatacatatgagcgataaaattattcacctgactgacgacagttttgacacggatgtactcaaagcggacggggcgatcctcgtcgatttctgggcagagtggtgcggtccgtgcaaaatgatcgccccgattctggatgaaatcgctgacgaatatcagggcaaactgaccgttgcaaaactgaacatcgatcaaaaccctggcactgcgccgaaatatggcatccgtggtatcccgactctgctgctgttcaaaaacggtgaagtggcggcaaccaaagtgggtgcactgtctaaaggtcagttgaaagagttcctcgacgctaacctggccggttctggttctggccatatgcaccatcatcatcatcattcttctggtgagaatctttattttcagggcgggtccgggttgaatgatattttcgaagcacagaaaattgaatggcatgaggccatgggcaacgatgtggttacctacaccaccctgatcgatggtctggcgaaagcgggccgtctggaggaagcgctgcagctgttccaagagatgaaggaaaaaggtgtgaagccggacgtggttacctataacaccctgattgacggtctggcgaaggctggccgtctggaggaagcgctgcaactgtttcaggaaatgaaggaaaaaggcgttaaaccggatgttgtgacctataccaccctgattgatggtctggcgaaggccggccgtctggaggaagcgctgtgcctgttccaggagatgaaagagaaaggtgtgaagccgaatgttgtgacctacaataccctgatcgacggtctggcgaaggcaggccgtctggaggaagcgctgcagttatttcaggagatgaaagaaaaaggcgttaagccggatgttgttacctacaacactttaattgacggtctggcgaaagctggccgtctggaggaagcgctgcagctatttcaggagatgaaggagaaaggtgtgaaaccgagcgtggtgacctacaacactttaatcgatggtctggcgaaggcgggccgtctggaggaagcgctgcagctttttcaggagatgaaggaaaaaggcgttaagccgagcgtggttacctacaacactcttattgatggtctggcgaaagctggccgtctggaggaagcgctgcagctctttcaggagatgaaggaaaaaggtgtgaaaccggatgttgttacctataatactttgatagatggtctggcgaaggcgggccgtctggaggaagcgctgcagctattccaggaaatgaaagagaaaggcgttaaaccggacgttgtgacctacaccactttaattgatggtctggcgaaagccggccgtctggaggaagcgctgcagctattccaagaaatgaaagaaaaaggtgtgaagcctaacgtggttacctataccactttaatcgacggtctggcgaaagccggccgtctggaggaagcgctgcagctattccaagagatgaaagagaaaggcgttaagccgaatgtggtgacctacaacacattaattgatggtctggcgaaagccggccgtctggaggaagcgctgtgtctgttccaggagatgaaggaaaaaggtgtgaagccgagcgtggtgacctataatactttgatcgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagaaatgaaagaaaaaggcgttaagccgagcgtggtaacctataccactcttatcgacggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaagaaaaaggtgtgaaacctaatgttgtgacctacaacactcttatcgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaggagaaaggcgttaagcctgatgtggttacctataatactctcatcgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaggaaaaaggtgtgaagccagatgttgtgacctataatactctaatagatggtctggcgaaggcgggccgtctggaggaagcgctgcagctatttcaagagatgaaggaaaaaggtgttaaaccggatgtggttacctacaacacacttattgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaggaaaagggtgtgaaaccgagcgttgtgaccaataacaccctgaaggacggcgcgagcaaagcgggctaagaattcgagctccgtcgacaagcttgcggccgcactcgagcaccaccaccaccaccactgagatccggctgctaacaaagcccgaaaggaagctgagttggctgctgccaccgctgagcaataactagcataaccccttggggcctctaaacgggtcttgaggggttttttgctgaaaggaggaactatatccggat
#> 2 tggcgaatgggacgcgccctgtagcggcgcattaagcgcggcgggtgtggtggttacgcgcagcgtgaccgctacacttgccagcgccctagcgcccgctcctttcgctttcttcccttcctttctcgccacgttcgccggctttccccgtcaagctctaaatcgggggctccctttagggttccgatttagtgctttacggcacctcgaccccaaaaaacttgattagggtgatggttcacgtagtgggccatcgccctgatagacggtttttcgccctttgacgttggagtccacgttctttaatagtggactcttgttccaaactggaacaacactcaaccctatctcggtctattcttttgatttataagggattttgccgatttcggcctattggttaaaaaatgagctgatttaacaaaaatttaacgcgaattttaacaaaatattaacgtttacaatttcaggtggcacttttcggggaaatgtgcgcggaacccctatttgtttatttttctaaatacattcaaatatgtatccgctcatgagacaataaccctgataaatgcttcaataatattgaaaaaggaagagtatgagtattcaacatttccgtgtcgcccttattcccttttttgcggcattttgccttcctgtttttgctcacccagaaacgctggtgaaagtaaaagatgctgaagatcagttgggtgcacgagtgggttacatcgaactggatctcaacagcggtaagatccttgagagttttcgccccgaagaacgttttccaatgatgagcacttttaaagttctgctatgtggcgcggtattatcccgtattgacgccgggcaagagcaactcggtcgccgcatacactattctcagaatgacttggttgagtactcaccagtcacagaaaagcatcttacggatggcatgacagtaagagaattatgcagtgctgccataaccatgagtgataacactgcggccaacttacttctgacaacgatcggaggaccgaaggagctaaccgcttttttgcacaacatgggggatcatgtaactcgccttgatcgttgggaaccggagctgaatgaagccataccaaacgacgagcgtgacaccacgatgcctgcagcaatggcaacaacgttgcgcaaactattaactggcgaactacttactctagcttcccggcaacaattaatagactggatggaggcggataaagttgcaggaccacttctgcgctcggcccttccggctggctggtttattgctgataaatctggagccggtgagcgtgggtctcgcggtatcattgcagcactggggccagatggtaagccctcccgtatcgtagttatctacacgacggggagtcaggcaactatggatgaacgaaatagacagatcgctgagataggtgcctcactgattaagcattggtaactgtcagaccaagtttactcatatatactttagattgatttaaaacttcatttttaatttaaaaggatctaggtgaagatcctttttgataatctcatgaccaaaatcccttaacgtgagttttcgttccactgagcgtcagaccccgtagaaaagatcaaaggatcttcttgagatcctttttttctgcgcgtaatctgctgcttgcaaacaaaaaaaccaccgctaccagcggtggtttgtttgccggatcaagagctaccaactctttttccgaaggtaactggcttcagcagagcgcagataccaaatactgtccttctagtgtagccgtagttaggccaccacttcaagaactctgtagcaccgcctacatacctcgctctgctaatcctgttaccagtggctgctgccagtggcgataagtcgtgtcttaccgggttggactcaagacgatagttaccggataaggcgcagcggtcgggctgaacggggggttcgtgcacacagcccagcttggagcgaacgacctacaccgaactgagatacctacagcgtgagctatgagaaagcgccacgcttcccgaagggagaaaggcggacaggtatccggtaagcggcagggtcggaacaggagagcgcacgagggagcttccagggggaaacgcctggtatctttatagtcctgtcgggtttcgccacctctgacttgagcgtcgatttttgtgatgctcgtcaggggggcggagcctatggaaaaacgccagcaacgcggcctttttacggttcctggccttttgctggccttttgctcacatgttctttcctgcgttatcccctgattctgtggataaccgtattaccgcctttgagtgagctgataccgctcgccgcagccgaacgaccgagcgcagcgagtcagtgagcgaggaagcggaagagcgcctgatgcggtattttctccttacgcatctgtgcggtatttcacaccgcatatatggtgcactctcagtacaatctgctctgatgccgcatagttaagccagtatacactccgctatcgctacgtgactgggtcatggctgcgccccgacacccgccaacacccgctgacgcgccctgacgggcttgtctgctcccggcatccgcttacagacaagctgtgaccgtctccgggagctgcatgtgtcagaggttttcaccgtcatcaccgaaacgcgcgaggcagctgcggtaaagctcatcagcgtggtcgtgaagcgattcacagatgtctgcctgttcatccgcgtccagctcgttgagtttctccagaagcgttaatgtctggcttctgataaagcgggccatgttaagggcggttttttcctgtttggtcactgatgcctccgtgtaagggggatttctgttcatgggggtaatgataccgatgaaacgagagaggatgctcacgatacgggttactgatgatgaacatgcccggttactggaacgttgtgagggtaaacaactggcggtatggatgcggcgggaccagagaaaaatcactcagggtcaatgccagcgcttcgttaatacagatgtaggtgttccacagggtagccagcagcatcctgcgatgcagatccggaacataatggtgcagggcgctgacttccgcgtttccagactttacgaaacacggaaaccgaagaccattcatgttgttgctcaggtcgcagacgttttgcagcagcagtcgcttcacgttcgctcgcgtatcggtgattcattctgctaaccagtaaggcaaccccgccagcctagccgggtcctcaacgacaggagcacgatcatgcgcacccgtggggccgccatgccggcgataatggcctgcttctcgccgaaacgtttggtggcgggaccagtgacgaaggcttgagcgagggcgtgcaagattccgaataccgcaagcgacaggccgatcatcgtcgcgctccagcgaaagcggtcctcgccgaaaatgacccagagcgctgccggcacctgtcctacgagttgcatgataaagaagacagtcataagtgcggcgacgatagtcatgccccgcgcccaccggaaggagctgactgggttgaaggctctcaagggcatcggtcgagatcccggtgcctaatgagtgagctaacttacattaattgcgttgcgctcactgcccgctttccagtcgggaaacctgtcgtgccagctgcattaatgaatcggccaacgcgcggggagaggcggtttgcgtattgggcgccagggtggtttttcttttcaccagtgagacgggcaacagctgattgcccttcaccgcctggccctgagagagttgcagcaagcggtccacgctggtttgccccagcaggcgaaaatcctgtttgatggtggttaacggcgggatataacatgagctgtcttcggtatcgtcgtatcccactaccgagatgtccgcaccaacgcgcagcccggactcggtaatggcgcgcattgcgcccagcgccatctgatcgttggcaaccagcatcgcagtgggaacgatgccctcattcagcatttgcatggtttgttgaaaaccggacatggcactccagtcgccttcccgttccgctatcggctgaatttgattgcgagtgagatatttatgccagccagccagacgcagacgcgccgagacagaacttaatgggcccgctaacagcgcgatttgctggtgacccaatgcgaccagatgctccacgcccagtcgcgtaccgtcttcatgggagaaaataatactgttgatgggtgtctggtcagagacatcaagaaataacgccggaacattagtgcaggcagcttccacagcaatggcatcctggtcatccagcggatagttaatgatcagcccactgacgcgttgcgcgagaagattgtgcaccgccgctttacaggcttcgacgccgcttcgttctaccatcgacaccaccacgctggcacccagttgatcggcgcgagatttaatcgccgcgacaatttgcgacggcgcgtgcagggccagactggaggtggcaacgccaatcagcaacgactgtttgcccgccagttgttgtgccacgcggttgggaatgtaattcagctccgccatcgccgcttccactttttcccgcgttttcgcagaaacgtggctggcctggttcaccacgcgggaaacggtctgataagagacaccggcatactctgcgacatcgtataacgttactggtttcacattcaccaccctgaattgactctcttccgggcgctatcatgccataccgcgaaaggttttgcgccattcgatggtgtccgggatctcgacgctctcccttatgcgactcctgcattaggaagcagcccagtagtaggttgaggccgttgagcaccgccgccgcaaggaatggtgcatgcaaggagatggcgcccaacagtcccccggccacggggcctgccaccatacccacgccgaaacaagcgctcatgagcccgaagtggcgagcccgatcttccccatcggtgatgtcggcgatataggcgccagcaaccgcacctgtggcgccggtgatgccggccacgatgcgtccggcgtagaggatcgagatcgatctcgatcccgcgaaattaatacgactcactataggggaattgtgagcggataacaattcccctctagaaataattttgtttaactttaagaaggagatatacatatgagcgataaaattattcacctgactgacgacagttttgacacggatgtactcaaagcggacggggcgatcctcgtcgatttctgggcagagtggtgcggtccgtgcaaaatgatcgccccgattctggatgaaatcgctgacgaatatcagggcaaactgaccgttgcaaaactgaacatcgatcaaaaccctggcactgcgccgaaatatggcatccgtggtatcccgactctgctgctgttcaaaaacggtgaagtggcggcaaccaaagtgggtgcactgtctaaaggtcagttgaaagagttcctcgacgctaacctggccggttctggttctggccatatgcaccatcatcatcatcattcttctggtgagaatctttattttcagggcgggtccgggttgaatgatattttcgaagcacagaaaattgaatggcatgaggccatgggcaacgatgtggttacctacaccaccctgatcgatggtctggcgaaagcgggccgtctggaggaagcgctgcagctgttccaagagatgaaggaaaaaggtgtgaagccggacgtggttacctataacaccctgattgacggtctggcgaaggctggccgtctggaggaagcgctgcaactgtttcaggaaatgaaggaaaaaggcgttaaaccggatgttgtgacctataccaccctgattgatggtctggcgaaggccggccgtctggaggaagcgctgtgcctgttccaggagatgaaagagaaaggtgtgaagccgaatgttgtgacctacaataccctgatcgacggtctggcgaaggcaggccgtctggaggaagcgctgcagttatttcaggagatgaaagaaaaaggcgttaagccggatgttgttacctacaacactttaattgacggtctggcgaaagctggccgtctggaggaagcgctgcagctatttcaggagatgaaggagaaaggtgtgaaaccgagcgtggtgacctacaacactttaatcgatggtctggcgaaggcgggccgtctggaggaagcgctgcagctttttcaggagatgaaggaaaaaggcgttaagccgagcgtggttacctacaacactcttattgatggtctggcgaaagctggccgtctggaggaagcgctgcagctctttcaggagatgaaggaaaaaggtgtgaaaccggatgttgttacctataatactttgatagatggtctggcgaaggcgggccgtctggaggaagcgctgcagctattccaggaaatgaaagagaaaggcgttaaaccggacgttgtgacctacaccactttaattgatggtctggcgaaagccggccgtctggaggaagcgctgcagctattccaagaaatgaaagaaaaaggtgtgaagcctaacgtggttacctataccactttaatcgacggtctggcgaaagccggccgtctggaggaagcgctgcagctattccaagagatgaaagagaaaggcgttaagccgaatgtggtgacctacaacacattaattgatggtctggcgaaagccggccgtctggaggaagcgctgtgtctgttccaggagatgaaggaaaaaggtgtgaagccgagcgtggtgacctataatactttgatcgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagaaatgaaagaaaaaggcgttaagccgagcgtggtaacctataccactcttatcgacggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaagaaaaaggtgtgaaacctaatgttgtgacctacaacactcttatcgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaggagaaaggcgttaagcctgatgtggttacctataatactctcatcgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaggaaaaaggtgtgaagccagatgttgtgacctataatactctaatagatggtctggcgaaggcgggccgtctggaggaagcgctgcagctatttcaagagatgaaggaaaaaggtgttaaaccggatgtggttacctacaacacacttattgatggtctggcgaaagcaggccgtctggaggaagcgctgcagctatttcaagagatgaaggaaaagggtgtgaaaccgagcgttgtgaccaataacaccctgaaggacggcgcgagcaaagcgggctaagaattcgagctccgtcgacaagcttgcggccgcactcgagcaccaccaccaccaccactgagatccggctgctaacaaagcccgaaaggaagctgagttggctgctgccaccgctgagcaataactagcataaccccttggggcctctaaacgggtcttgaggggttttttgctgaaaggaggaactatatccggat